flow-tests/pyproject.toml
[project]
name = "redisearch-disk-tests"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "gevent<=24.11.1",
    "packaging<=24.2",
    "redis>=5.2.1,<7.0.0",
    "RLTest>=0.7.18,<0.8.0",
    "pytest-bdd>=8.1.0,<9.0.0",
    "pytest-xdist>=3.0.0,<4.0.0",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["steps*", "features*"]

[dependency-groups]
dev = []

